+++ tests/test_module_generator.py
@@ -26,12 +26,13 @@
 from wishbone.module.generator import Generator
 
 from wishbone.actor import ActorConfig
+from wishbone.queue import QueuePool
 from wishbone.utils.test import getter
 
 
 def test_module_generator_basic():
 
-    actor_config = ActorConfig('generator', 100, 1, {}, "")
+    actor_config = ActorConfig('generator', QueuePool())
     test_event = Generator(actor_config, payload="test")
 
     test_event.pool.queue.outbox.disableFallThrough()
@@ -43,7 +44,7 @@ def test_module_generator_basic():
 
 def test_module_generator_dict():
 
-    actor_config = ActorConfig('generator', 100, 1, {}, "")
+    actor_config = ActorConfig('generator', QueuePool())
     test_event = Generator(actor_config, payload={"one": 1})
 
     test_event.pool.queue.outbox.disableFallThrough()