+++ tests/test_module_switch.py
@@ -25,12 +25,13 @@
 from wishbone.event import Event
 from wishbone.module.switch import Switch
 from wishbone.actor import ActorConfig
+from wishbone.queue import QueuePool
 from wishbone.utils.test import getter
 
 
 def test_module_switch_default():
 
-    actor_config = ActorConfig('switch', 100, 1, {}, "")
+    actor_config = ActorConfig('switch', QueuePool())
 
     switch = Switch(actor_config, outgoing="one")
     switch.pool.queue.inbox.disableFallThrough()
@@ -52,7 +53,7 @@ def test_module_switch_default():
 
 def test_module_switch_event():
 
-    actor_config = ActorConfig('switch', 100, 1, {}, "", disable_exception_handling=True)
+    actor_config = ActorConfig('switch', QueuePool(), disable_exception_handling=True)
 
     switch = Switch(actor_config, outgoing="one")
     switch.pool.queue.inbox.disableFallThrough()