+++ wishbone/actorconfig.py
@@ -25,7 +25,7 @@
 
 class ActorConfig(object):
 
-    '''
+    """
     A configuration object pass to a Wishbone actor.
 
     This is a simple object which holds a set of attributes (with some sane
@@ -33,7 +33,7 @@ class ActorConfig(object):
 
     Attributes:
         name (str): The name identifying the actor instance.
-        size (int): The size of the Actor instance's queues.
+        queue_pool (QueuePool): The QueuePool instance.
         frequency (int): The time in seconds to generate metrics.
         template_functions (dict): A dictionary of template functions.
         description (str): A short free form discription of the actor instance.
@@ -43,16 +43,25 @@ class ActorConfig(object):
         io_event (bool): When ``True`` Input and Output modules know to expect or emit serialzed wishbone events.
         identification (str): A name assigned to the Wishbone instance, useful for the module to know such as logging.
         disable_exception_handling (bool): If True, exception handling is disabled. Usefull for testing
-    '''
+    """
 
-    def __init__(self, name, size=100, frequency=10, template_functions={}, description=None, module_functions={},
-                 protocol=None, io_event=False,
-                 identification="wishbone",
-                 disable_exception_handling=False):
-        '''
+    def __init__(
+        self,
+        name,
+        queue_pool,
+        frequency=10,
+        template_functions={},
+        description=None,
+        module_functions={},
+        protocol=None,
+        io_event=False,
+        identification="wishbone",
+        disable_exception_handling=False,
+    ):
+        """
         Args:
             name (str): The name identifying the actor instance.
-            size (int): The size of the Actor instance's queues.
+            queue_pool (QueuePool): The QueuePool instance.
             frequency (int): The time in seconds to generate metrics.
             template_functions (dict): A dictionary of template functions.
             description (str): A short free form discription of the actor instance.
@@ -62,9 +71,8 @@ class ActorConfig(object):
             io_event (bool): When ``True`` Input and Output modules know to expect or emit serialzed wishbone events.
             identification (str): A name assigned to the Wishbone instance, useful for the module to know such as logging.
             disable_exception_handling (bool): If True, exception handling is disabled. Usefull for testing
-        '''
+        """
         self.name = name
-        self.size = size
         self.frequency = frequency
         self.template_functions = template_functions
         self.description = description
@@ -73,3 +81,4 @@ class ActorConfig(object):
         self.io_event = io_event
         self.identification = identification
         self.disable_exception_handling = disable_exception_handling
+        self.queue_pool = queue_pool