+++ release_notes.rst
@@ -1,6 +1,18 @@
-Version 3.1.4
+Version 3.1.5
 =============
 
-Bugfixes:
+Changes:
 
-    - Fix bug where setting encoder replace decoder.
+    - Introduced `wishbone.queue.QueuePool` object to simplify queue
+      management outside of the Actor in preparation of using different queue
+      Types.
+
+    - Introduced `wishbone.queue.MemoryQueue` the default Wishbone queue type.
+
+    - Modified `wishbone.actor.Actorconfig` object signature to pass the
+      `wishbone.queue.QueuePool` object instance to it.
+
+    - When connecting two queues we are no long overwriting one instance with
+      the other but we are actually shoveling messages from one queue to the
+      other using a greenthread.
+